järjestelty komennot tiedostoihin
