+++ src/org/kaivos/röda/commands/ErrprintPopulator.java
@@ -1,0 +1,30 @@
+package org.kaivos.röda.commands;
+
+import java.util.Arrays;
+
+import org.kaivos.röda.Interpreter.RödaScope;
+import org.kaivos.röda.Parser.Parameter;
+import org.kaivos.röda.RödaValue;
+import org.kaivos.röda.type.RödaNativeFunction;
+import org.kaivos.röda.type.RödaString;
+
+public final class ErrprintPopulator {
+
+	private ErrprintPopulator() {}
+
+	public static void populateErrprint(RödaScope S) {
+		S.setLocal("errprint", RödaNativeFunction.of("errprint", (typeargs, args, scope, in, out) -> {
+					if (args.isEmpty()) {
+						while (true) {
+							RödaValue input = in.pull();
+							if (input == null) break;
+							System.err.print(input.str());
+						}
+					}
+					else for (RödaValue value : args) {
+						System.err.print(value.str());
+						out.push(RödaString.of("\n"));
+					}
+				}, Arrays.asList(new Parameter("values", false)), true));
+	}
+}
