+++ src/org/kaivos/röda/commands/GetenvPopulator.java
@@ -1,0 +1,21 @@
+package org.kaivos.röda.commands;
+
+import static org.kaivos.röda.RödaValue.STRING;
+
+import java.util.Arrays;
+
+import org.kaivos.röda.Interpreter.RödaScope;
+import org.kaivos.röda.Parser.Parameter;
+import org.kaivos.röda.type.RödaNativeFunction;
+import org.kaivos.röda.type.RödaString;
+
+public final class GetenvPopulator {
+
+	private GetenvPopulator() {}
+
+	public static void populateGetenv(RödaScope S) {
+		S.setLocal("getenv", RödaNativeFunction.of("getenv", (typeargs, args, scope, in, out) -> {
+				        out.push(RödaString.of(System.getenv(args.get(0).str())));
+				}, Arrays.asList(new Parameter("name", false, STRING)), false));
+	}
+}
