+++ src/org/kaivos/röda/commands/JsonPopulator.java
@@ -1,0 +1,84 @@
+package org.kaivos.röda.commands;
+
+import static java.util.stream.Collectors.toList;
+import static org.kaivos.röda.Interpreter.argumentOverflow;
+import static org.kaivos.röda.Interpreter.checkString;
+
+import java.util.Arrays;
+import java.util.function.Consumer;
+
+import org.kaivos.röda.Interpreter.RödaScope;
+import org.kaivos.röda.JSON;
+import org.kaivos.röda.JSON.JSONElement;
+import org.kaivos.röda.JSON.JSONInteger;
+import org.kaivos.röda.JSON.JSONList;
+import org.kaivos.röda.JSON.JSONMap;
+import org.kaivos.röda.JSON.JSONString;
+import org.kaivos.röda.Parser.Parameter;
+import org.kaivos.röda.RödaValue;
+import org.kaivos.röda.type.RödaInteger;
+import org.kaivos.röda.type.RödaList;
+import org.kaivos.röda.type.RödaNativeFunction;
+import org.kaivos.röda.type.RödaString;
+
+public final class JsonPopulator {
+
+	private JsonPopulator() {}
+
+	public static void populateJson(RödaScope S) {
+		S.setLocal("json", RödaNativeFunction.of("json", (typeargs, args, scope, in, out) -> {
+					Consumer<String> handler = code -> {
+						JSONElement root = JSON.parseJSON(code);
+						// rekursiivinen ulostulo
+						// apuluokka rekursion mahdollistamiseksi
+						class R<I> { I i; }
+						R<java.util.function.Function<JSONElement, RödaValue>>
+						makeRöda = new R<>();
+						makeRöda.i = json -> {
+							RödaValue elementName = RödaString.of(json.getElementName());
+							RödaValue value;
+							if (json instanceof JSONInteger) {
+								value = RödaInteger.of(((JSONInteger) json).getValue());
+							}
+							else if (json instanceof JSONString) {
+								value = RödaString.of(((JSONString) json).getValue());
+							}
+							else if (json instanceof JSONList) {
+								value = RödaList.of(((JSONList) json).getElements()
+										      .stream()
+										      .map(j -> makeRöda.i.apply(j))
+										      .collect(toList()));
+							}
+							else if (json instanceof JSONMap) {
+								value = RödaList.of(((JSONMap) json).getElements().entrySet()
+										      .stream()
+										      .map(e -> RödaList.of(RödaString.of(e.getKey().getKey()),
+													      makeRöda.i.apply(e.getValue())))
+										      .collect(toList()));
+							}
+							else {
+								value = RödaString.of(json.toString());
+							}
+							return RödaList.of(elementName, value);
+						};
+						out.push(makeRöda.i.apply(root));
+						
+					
+					};
+					if (args.size() > 1) argumentOverflow("json", 1, args.size());
+					else if (args.size() == 1) {
+						RödaValue arg = args.get(0);
+						checkString("json", arg);
+						String code = arg.str();
+						handler.accept(code);
+					}
+					else {
+						RödaValue value;
+						while ((value = in.pull()) != null) {
+							String code = value.str();
+							handler.accept(code);
+						}
+					}
+				}, Arrays.asList(new Parameter("flags_and_code", false)), true));
+	}
+}
