+++ src/org/kaivos/röda/commands/NamePopulator.java
@@ -1,0 +1,28 @@
+package org.kaivos.röda.commands;
+
+import static org.kaivos.röda.Interpreter.error;
+
+import java.util.Arrays;
+
+import org.kaivos.röda.Interpreter.RödaScope;
+import org.kaivos.röda.Parser.Parameter;
+import org.kaivos.röda.RödaValue;
+import org.kaivos.röda.type.RödaNativeFunction;
+import org.kaivos.röda.type.RödaString;
+
+public final class NamePopulator {
+	
+	private NamePopulator() {}
+
+	public static void populateName(RödaScope S) {
+		S.setLocal("name", RödaNativeFunction.of("name", (typeargs, args, scope, in, out) -> {
+					for (RödaValue value : args) {
+						if (!value.is(RödaValue.REFERENCE))
+							error("invalid argument for undefine: "
+							      + "only references accepted");
+	
+						out.push(RödaString.of(value.target()));
+					}
+				}, Arrays.asList(new Parameter("variables", true)), true));
+	}
+}
