+++ src/org/kaivos/röda/commands/StrsizePopulator.java
@@ -1,0 +1,41 @@
+package org.kaivos.röda.commands;
+
+import static org.kaivos.röda.Interpreter.checkString;
+import static org.kaivos.röda.RödaValue.STRING;
+
+import java.nio.charset.Charset;
+import java.nio.charset.StandardCharsets;
+import java.util.Arrays;
+import java.util.function.Consumer;
+
+import org.kaivos.röda.Interpreter.RödaScope;
+import org.kaivos.röda.Parser.Parameter;
+import org.kaivos.röda.RödaValue;
+import org.kaivos.röda.type.RödaInteger;
+import org.kaivos.röda.type.RödaNativeFunction;
+
+public final class StrsizePopulator {
+
+	private StrsizePopulator() {}
+
+	public static void populateStrsize(RödaScope S) {
+		S.setLocal("strsize", RödaNativeFunction.of("strsize", (typeargs, args, scope, in, out) -> {
+					Charset chrset = StandardCharsets.UTF_8;
+					Consumer<RödaValue> convert = v -> {
+							checkString("strsize", v);
+							out.push(RödaInteger.of(v.str().getBytes(chrset).length));
+					};
+				        if (args.size() > 0) {
+						for (RödaValue v : args) {
+							convert.accept(v);
+						}
+					} else {
+						while (true) {
+							RödaValue v = in.pull();
+							if (v == null) break;
+							convert.accept(v);
+						}
+					}
+				}, Arrays.asList(new Parameter("strings", false, STRING)), true));
+	}
+}
